refactor: extract request classes to `request` package
Extract request-related classes from `AuthorizationCheckBehavior` to dedicated package:
- AuthorizationRequestMetadata: immutable request data
- AuthorizationRequest: builder class

(cherry picked from commit a8e1577e2a35a31d1f7ef53e1d41df07a527fd40)
